replaced userId by profileId

+++ src/folders/folder.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, Logger, NotFoundException } from '@nestjs/common';
+import { BadGatewayException, Injectable, Logger, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from 'src/dynamo-db/decorators/inject-model.decorator';
 import { Folder } from './folder.model';
 import { DynamoDbRepository } from 'src/dynamo-db/dynamo-db.repository';
@@ -17,14 +17,23 @@ export class FolderService {
     private readonly profileService: ProfileService
   ) { }
 
-  async findUserFolderById(id: string, userId: string) {
+  async findUserFolderByIdAndUserId(id: string, userId: string) {
 
+    const profile = await this.profileService.findProfileByUserId(userId);
+    if (!profile) {
+      throw new NotFoundException('Profile not found');
+    }
+
+    return this.findUserFolderByIdAndProfileId(id, profile.id);
+  }
+
+  async findUserFolderByIdAndProfileId(id: string, profileId: string) {
     const count = await this.photoService
-      .getTotalPhotosByFolderId(id, userId);
+      .getTotalPhotosByFolderId(id, profileId);
 
     return this.folderRepository
       .findOneByFilter<Folder>({
-        match: { id, userId }
+        match: { id, profileId: profileId },
       })
       .then((data) => ({ ...data, totalPhotos: count || 0 }))
       .catch((err) => {
@@ -34,39 +43,66 @@ export class FolderService {
   }
 
   async findAllByUserId(userId: string) {
-    const folders = await this.folderRepository.readByFilter<Folder>({ match: { userId } });
+    const profile = await this.profileService.findProfileByUserId(userId);
+    if (!profile) {
+      throw new NotFoundException('Profile not found');
+    }
+
+    const folders = await this.folderRepository.readByFilter<Folder>({ match: { profileId: profile.id } });
     return Promise.all(folders.map(async folder => {
-      const count = await this.photoService.getTotalPhotosByFolderId(folder.id, userId);
+      const count = await this.photoService.getTotalPhotosByFolderId(folder.id, profile.id);
       return { ...folder, totalPhotos: count };
     }));
   }
 
-  async createFolder(data: Partial<Folder>, userName: string) {
-    const profile = await this.profileService.findProfileByUserId(data.userId);
+  async findAllByProfileId(profileId: string) {
+    const folders = await this.folderRepository.readByFilter<Folder>({ match: { profileId: profileId } });
+    return Promise.all(folders.map(async folder => {
+      const count = await this.photoService.getTotalPhotosByFolderId(folder.id, profileId);
+      return { ...folder, totalPhotos: count };
+    }));
+  }
+
+  async createFolder(data: Partial<Folder>, userId: string) {
+    const profile = await this.profileService.findProfileByUserId(userId) || await this.profileService.createProfile(userId, 'unknown');
     if (!profile) {
-      await this.profileService.createProfile(data.userId, userName);
+      throw new NotFoundException('Profile not found');
     }
 
-    const userFolders = await this.findAllByUserId(data.userId);
+    const userFolders = await this.findAllByProfileId(profile.id);
     if (userFolders.length >= 10) {
-      throw new Error(
+      throw new BadGatewayException(
         `You can't create more than 10 folders. You have ${userFolders.length}`,
       );
     }
-    return this.folderRepository.create(data);
+    return this.folderRepository.create({ ...data, profileId: profile.id });
   }
 
   async updateFolder(id: string, data: FolderInputDto, userId: string) {
-    return this.folderRepository.update(id, { ...data, userId });
+    const profile = await this.profileService.findProfileByUserId(userId);
+    if (!profile) {
+      throw new NotFoundException('Profile not found');
+    }
+
+    const userFolder = await this.findUserFolderByIdAndProfileId(id, profile?.id);
+    if (!userFolder) {
+      throw new NotFoundException(`Folder with id ${id} not found`);
+    }
+
+    return this.folderRepository.update(id, { ...data, profileId: profile.id });
   }
 
   async removeFolder(id: string, userId: string) {
-    const userFolder = await this.findUserFolderById(id, userId);
+    const profile = await this.profileService.findProfileByUserId(userId);
+    if (!profile) {
+      throw new NotFoundException('Profile not found');
+    }
+    const userFolder = await this.findUserFolderByIdAndProfileId(id, profile?.id);
     if (!userFolder) {
       throw new NotFoundException(`Folder with id ${id} not found`);
     }
     //delete all photos from folder
-    await this.photoService.removePhotosByFolderId(id, userId);
+    await this.photoService.removePhotosByFolderId(id, profile.id);
     return this.folderRepository.remove(id);
   }
 }
